+++ valhalla/baldr/directededge.h
@@ -565,6 +565,17 @@ public:
   }
 
   /**
+   * Evaluates a basic set of conditions to determine if this directed edge is a valid potential
+   * member of a shortcut. This is used while forming and resolving shortcuts.
+   * @return true if the edge is not a shortcut, not related to transit and not under construction
+   */
+  bool can_form_shortcut() const {
+    return !is_shortcut() && !bss_connection() && !start_restriction() && !end_restriction() &&
+           use() != Use::kTransitConnection && use() != Use::kEgressConnection &&
+           use() != Use::kPlatformConnection && use() != Use::kConstruction;
+  }
+
+  /**
    * Sets the specialized use type of this edge.
    * @param  use  Use of this edge.
    */
@@ -1074,6 +1085,29 @@ public:
    */
   uint32_t superseded() const {
     return superseded_;
+  }
+
+#ifdef _WIN32
+  // TODO: Workaround for missing strings.h on Windows. Replace with platform independent
+  //       std::countr_zero in C++20 (see https://en.cppreference.com/w/cpp/numeric/countr_zero).
+  int ffs(int mask) const {
+    if (0 == mask)
+      return 0;
+
+    int idx;
+    for (idx = 1; !(mask & 1); ++idx)
+      mask >>= 1;
+    return idx;
+  }
+#endif
+
+  /**
+   * Unlike superseded(), this does not return the raw mask but the shortcut index
+   * that was originally passed to set_superseded().
+   * @return  Returns the index of the set bit in the superseded mask.
+   */
+  uint32_t superseded_idx() const {
+    return ffs(superseded_);
   }
 
   /**